pyproject.toml
[project]
# Core project metadata (PEP 621)
name = "semantiq"
version = "0.1.0"
description = "Semantic Intelligence Benchmarking Platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [ { name = "SemantIQ Contributors" } ]
keywords = ["LLM", "benchmark", "evaluation", "semantics", "AI"]

# Core dependencies kept minimal yet functional for Phase 0
dependencies = [
  "pydantic>=2.4",
  "PyYAML>=6.0",
  "httpx>=0.24",
  "fastapi>=0.110",
  "uvicorn>=0.22",
  "jinja2>=3.1",
  "typer>=0.9"
]

[project.optional-dependencies]
providers = [
  "httpx>=0.24",
  "openai>=1.0.0"
]
yaml = [
  "PyYAML>=6.0"
]
dashboard = [
  "fastapi>=0.110",
  "uvicorn>=0.22",
  "jinja2>=3.1"
]
human_rater = [
  "fastapi>=0.110",
  "uvicorn>=0.22",
  "jinja2>=3.1"
]
all = [
  "PyYAML>=6.0",
  "httpx>=0.24",
  "openai>=1.0.0",
  "fastapi>=0.110",
  "uvicorn>=0.22",
  "jinja2>=3.1",
  "typer>=0.9"
]
cloud = [
  "sqlmodel>=0.0.14",
  "asyncpg>=0.29",
  "arq>=0.26",
  "redis>=5.0"
]
parquet = [
  "pandas>=2.0",
  "pyarrow>=16.0"
]
dev = [
  "pytest>=7.0",
  "ruff>=0.3",
  "black>=24.0"
]

[project.scripts]
semantiq = "semantiq.cli:app"

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/semantiq"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
